requirements.txt
autopep8==1.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
ffmpeg==1.4
Flask==2.2.2
idna==3.4
importlib-metadata==4.12.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
pycodestyle==2.9.1
pycountry==22.3.5
pycryptodome==3.15.0
PySocks==1.7.1
python-dotenv==0.21.0
python-slugify==6.1.2
requests==2.28.1
six==1.16.0
streamlink==5.0.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.12
websocket-client==1.4.1
Werkzeug==2.2.2
zipp==3.8.1
